get video info

+++ android/FFmpeg/library/src/main/cpp/video.h
@@ -59,7 +59,7 @@ public:
 
     double getDuration();
 
-    int getFrameRate();
+    double getFrameRate();
 
     void close();
 };
